Improve typing and type security
Changed typing of all speed, position, and orientation data to float in
order to smooth out jitters in the movement of objects. Also generally
made typing more explicit and replaced all type conversions with
explicit casts.

+++ code/include/asteroid.h
@@ -8,23 +8,23 @@ namespace ag {
 class Asteroid {
  public:
   Asteroid() {};
-  Asteroid(sf::Vector2i start_position);
+  Asteroid(sf::Vector2f start_position);
   ~Asteroid() {};
 
   void update_pos(const sf::Time dt);
   void reset_asteroid();
 
   sf::CircleShape sprite;
-  sf::Vector2i position;
+  sf::Vector2f position;
 
  private:
-  const float ASTEROID_SPEED = 100.0f;
+  const unsigned int ASTEROID_SPEED = 100u;
 
   void initialize_sprite_position();
   void initialize_sprite_graphics();
 
   sf::Vector2f velocity;
-  int orientation;
+  float orientation;
 };
 
 }
